deploy: build before deploy

+++ deploy.sh
@@ -1,8 +1,9 @@
 #!/usr/bin/env bash
 
+[ "$1" = "clean" ] && rm -rf public
+hugo
+
 hugo deploy --target=emstone.nvrsw.com
 exit $?
 
-rm -rf public
-hugo
 aws s3 sync public s3://emstone.nvrsw.com/ --delete
